+++ R/XenaHub-class.R
@@ -22,7 +22,8 @@ xena_default_hosts = function() {
       "https://toil.xenahubs.net",
       "https://pancanatlas.xenahubs.net",
       "https://xena.treehouse.gi.ucsc.edu",
-      "https://pcawg.xenahubs.net"
+      "https://pcawg.xenahubs.net",
+      "https://atacseq.xenahubs.net"
       )
 }
 
@@ -40,7 +41,7 @@ xena_default_hosts = function() {
 ##' for substitute this.
 ##' @param cohorts default is empty character vector, all cohorts will be returned.
 ##' @param datasets default is empty character vector, all datasets will be returned.
-##' @param hostName one to seven of \code{c("UCSC_Public", "TCGA", "GDC", "ICGC", "Toil", "PanCancer", "Treehouse", "PCAWG")}. This is
+##' @param hostName one to seven of \code{c("UCSC_Public", "TCGA", "GDC", "ICGC", "Toil", "PanCancer", "Treehouse", "PCAWG", "ATACseq")}. This is
 ##' a easier option for user than \code{hosts} option. Note, this option will overlap \code{hosts}.
 ##' @return a \code{XenaHub} object
 ##' @author Shixiang Wang <w_shixiang@163.com>
@@ -62,14 +63,15 @@ xena_default_hosts = function() {
 ##' samples(xe)   # get samples
 ##' }
 XenaHub = function(hosts=xena_default_hosts(), cohorts=character(),
-             datasets=character(), hostName=c("UCSC_Public", "TCGA", "GDC", "ICGC", "Toil", "PanCancer", "Treehouse", "PCAWG")){
+             datasets=character(), hostName=c("UCSC_Public", "TCGA", "GDC", "ICGC",
+                                              "Toil", "PanCancer", "Treehouse", "PCAWG","ATACseq")){
 
     stopifnot(is.character(hosts), is.character(cohorts),
               is.character(datasets))
 
     hostName = unique(hostName)
 
-    if(length(hostName) != 7 & all(hostName %in% c("UCSC_Public", "TCGA", "GDC", "ICGC", "Toil", "PanCancer", "Treehouse", "PCAWG")) ){
+    if(length(hostName) != 9 & all(hostName %in% c("UCSC_Public", "TCGA", "GDC", "ICGC", "Toil", "PanCancer", "Treehouse", "PCAWG", "ATACseq")) ){
       hostNames = data.frame(UCSC_Public="https://ucscpublic.xenahubs.net",
                              TCGA="https://tcga.xenahubs.net",
                              GDC="https://gdc.xenahubs.net",
@@ -78,6 +80,7 @@ XenaHub = function(hosts=xena_default_hosts(), cohorts=character(),
                              PanCancer="https://pancanatlas.xenahubs.net",
                              Treehouse="https://xena.treehouse.gi.ucsc.edu",
                              PCAWG="https://pcawg.xenahubs.net",
+                             ATACseq="https://atacseq.xenahubs.net",
                              stringsAsFactors = FALSE)
       hosts = as.character(hostNames[, hostName])
     }
@@ -160,7 +163,7 @@ XenaDataUpdate = function(saveTolocal=TRUE){
         resDF = rbind(resDF, res)
     }
 
-    XenaHostNames = c("UCSC_Public", "TCGA", "GDC", "ICGC", "Toil", "PanCancer", "Treehouse", "PCAWG")
+    XenaHostNames = c("UCSC_Public", "TCGA", "GDC", "ICGC", "Toil", "PanCancer", "Treehouse", "PCAWG", "ATACseq")
     names(XenaHostNames) = c("https://ucscpublic.xenahubs.net",
                              "https://tcga.xenahubs.net",
                              "https://gdc.xenahubs.net",
@@ -168,7 +171,8 @@ XenaDataUpdate = function(saveTolocal=TRUE){
                              "https://toil.xenahubs.net",
                              "https://pancanatlas.xenahubs.net",
                              "https://xena.treehouse.gi.ucsc.edu",
-                             "https://pcawg.xenahubs.net")
+                             "https://pcawg.xenahubs.net",
+                             "https://atacseq.xenahubs.net")
 
     resDF$XenaHostNames = XenaHostNames[resDF$XenaHosts]
     XenaData = resDF[, c("XenaHosts", "XenaHostNames", "XenaCohorts", "XenaDatasets")]